backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
pytest>=7.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
